move centre_from_points_qs function to uk_geo_utils app
