+++ polling_stations/apps/addressbase/helpers.py
@@ -8,21 +8,6 @@ from pollingstations.helpers import format_postcode_no_space
 from addressbase.models import Address
 
 
-def centre_from_points_qs(qs):
-    if not qs:
-        return None
-
-    if len(qs) == 1:
-        return qs[0].location
-
-    base_point = qs[0].location
-    poly = base_point.union(qs[1].location)
-    for m in qs:
-        poly = poly.union(m.location)
-
-    return poly.centroid
-
-
 AddressTuple = namedtuple('Address', [
     'address',
     'postcode',
