+++ polling_stations/apps/uk_geo_utils/fixtures/test_centroid.json
@@ -1,0 +1,29 @@
+[
+    {
+        "pk": "1",
+        "fields": {
+            "postcode": "KW15 88TF",
+            "address": "In the middle of the field",
+            "location": "SRID=4326;POINT (0.7208994610313337 52.0965737442737833)"
+        },
+        "model": "uk_geo_utils.address"
+    },
+    {
+        "pk": "2",
+        "fields": {
+            "postcode": "KW15 88TF",
+            "address": "The Cott",
+            "location": "SRID=4326;POINT (0.7225083308451804 52.0975051994085803)"
+        },
+        "model": "uk_geo_utils.address"
+    },
+    {
+        "pk": "3",
+        "fields": {
+            "postcode": "KW15 88LM",
+            "address": "Sym's Caravan",
+            "location": "SRID=4326;POINT (0.7152352684829660 52.0954911675566947)"
+        },
+        "model": "uk_geo_utils.address"
+    }
+]
